+++ packages/backend-proxy-middleware/src/base/proxy.ts
@@ -286,6 +286,9 @@ export async function generateProxyMiddlewareOptions(
     proxyOptions.changeOrigin = true;
     proxyOptions.logProvider = () => logger;
 
+    // always set the target to the url provided in yaml
+    proxyOptions.target = backend.url;
+
     // overwrite url if running in AppStudio
     if (isAppStudio()) {
         const destBackend = backend as DestinationBackendConfig;
@@ -296,7 +299,6 @@ export async function generateProxyMiddlewareOptions(
         }
     } else {
         const localBackend = backend as LocalBackendConfig;
-        proxyOptions.target = localBackend.url;
         // check if system credentials are stored in the store
         const systemStore = await getService<BackendSystem, BackendSystemKey>({ logger, entityName: 'system' });
         const system = await systemStore.read(
@@ -329,16 +331,16 @@ export async function generateProxyMiddlewareOptions(
         }
     }
 
+    if (!proxyOptions.target) {
+        throw new Error(`Unable to determine target from configuration:\n${JSON.stringify(backend, null, 2)}`);
+    }
+
     backend.proxy = getCorporateProxyServer(backend.proxy);
-    if (backend.proxy && !isHostExcludedFromProxy(proxyOptions.target as string)) {
+    if (backend.proxy && !isHostExcludedFromProxy(proxyOptions.target)) {
         proxyOptions.agent = new HttpsProxyAgent(backend.proxy);
     }
 
-    logger.info(
-        `Backend proxy created for ${proxyOptions.target ? proxyOptions.target : ''} ${
-            backend.path ? backend.path : ''
-        }`
-    );
+    logger.info(`Backend proxy created for ${proxyOptions.target} ${backend.path ? backend.path : ''}`);
     return proxyOptions;
 }