+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.6.8
+
+### Patch Changes
+
+-   fa232ad: Add optonal url property to the type DestinationBackendConfig
+
+## 0.6.7
+
+### Patch Changes
+
+-   7f339ca: Support using direct OData service URLs in BAS
+
 ## 0.6.6
 
 ### Patch Changes